Massive Refactor: Breaking the project into multiple files
- Moved all addresses, values, and utils to their own files
- Rewrote all effects as EffectHandler implementations
- Trimmed off all of the fat that was essentially rewriting what was
available in the SDK that we didn't understand or know about when we
first made this

This should make this project more maintainable and extensible, which is
exciting.

This does wipe out the existing menu we had and simplifies effect IDs.

We will need to have the menu file regenerated in the CC backend when
this gets merged in.

Effect specific changes:
- Consolidated TinyWeapon and GiantWeapon logic into one file
- Fixes bug in TinyWeapon/GiantWeapon where the weapon would be
  backwards during and after the effect
- Extend Backseat Driver to support individual form changes
- Consolidate Pro and Ez codes into one implementation.
- Fixed bugs in Pro and Ez Codes where values were mangled or overflowed